chore(tool): consolidate util package tests into per-source-file pairing
Third package in the test-consolidation pass for #1128 (tool/ex, then
tool/internal/ast, went first). tool/util had three orphaned test
files with no direct source-file pairing, and unlike the previous two
packages, their content wasn't a single-target merge -- each one is a
grab-bag spanning multiple source files:

  helpers_extra_test.go split four ways:
    TestCRC32, TestIsUnixAndIsWindows, TestNormalizePath -> sys_test.go
    TestQuoteGoflagsToken, TestNewFileScanner            -> go_test.go
    TestContextLogger                                    -> log_test.go
    TestAssertPasses, TestAssertType                     -> assert_test.go

  shared_paths_test.go -> shared_test.go (all four tests; every
    function under test -- GetOtelcWorkDir, GetBuildTempDir,
    GetMatchedRuleFile, GetAddedImportsPattern,
    GetAddedImportsFileForProcess, EncodeBuildFlags -- lives in
    shared.go)

  toolexec_dropin_test.go split two ways:
    TestDiscoverWorkDir          -> shared_test.go (DiscoverWorkDir is
                                     in shared.go)
    TestStripToolexecFromGoflags -> go_test.go (StripToolexecFromGoflags
                                     is in go.go)

This is a pure move: no test was added, removed, renamed, or edited.
The {name, pass/fail} set is identical before and after the merge.
